refactor model_registry.py to models
