+++ ocpmodels/models/__init__.py
@@ -2,3 +2,8 @@
 #
 # This source code is licensed under the MIT license found in the
 # LICENSE file in the root directory of this source tree.
+from __future__ import annotations
+
+from .model_registry import available_pretrained_models, model_name_to_local_file
+
+__all__ = ["available_pretrained_models", "model_name_to_local_file"]
